+++ src/main.ts
@@ -8,6 +8,7 @@ import { registerDataviewViewRenderers } from "./ux/dataview-views";
 import { registerExplorerActions } from "./ux/explorer-actions";
 import { registerLatestRetroSummaryRenderers } from "./ux/latest-retro-summary";
 import { refreshRegisteredLocaleLabels } from "./ux/locale-labels";
+import { registerManagedSectionRenderers } from "./ux/managed-sections";
 import { registerPropsControlRenderers } from "./ux/props-controls";
 import { registerRibbonActions } from "./ux/ribbon-actions";
 import { ParaZkSettingTab } from "./ux/settings";
@@ -35,6 +36,7 @@ export default class ParaZkPlugin extends Plugin {
     registerPropsControlRenderers(this);
     registerTaskRenderers(this);
     registerReferenceRenderers(this);
+    registerManagedSectionRenderers(this);
 
     this.addSettingTab(new ParaZkSettingTab(this));
     registerNativeCliHandlers(this);